requirements.txt
Django==4.2.4
django_embed_video==1.4.4
numpy==1.23.1
requests==2.28.1
whitenoise
mysqlclient==2.2.0
django-storages==1.13.2
azure-core==1.29.3
azure-storage-blob==12.17.0
